requirements.txt
numpy
opencv-python
matplotlib
pandas
ipykernel
keras_ocr
easyocr
tensorflow
deepface
streamlit
mysql-connector-python
sqlalchemy
pymysql